@@ -7,7 +7,7 @@ readme = "README.md"
 packages = [{ include = "SimpleLLMFunc", from = "." }]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<4.0"
+python = ">=3.11,<4.0"
 openai = ">=1.84.0,<2.0.0"
 pyyaml = ">=6.0.2,<7.0.0"
 pydantic = ">=2.11.5,<3.0.0"
